+++ src/components/Modal/BlightbaneModal/index.scss
@@ -1,45 +1,12 @@
-@import '../../styles/animations.scss';
-@import '../../styles/gradients.scss';
+@import '../../../styles/animations.scss';
+@import '../../../styles/gradients.scss';
 
 $button-padding-horizontal: 10px;
 $button-padding-vertical: 8px;
 
-@keyframes peekUp {
-  from {
-    transform: translateY(100%);
-    opacity: 0;
-  }
-  to {
-    transform: translateY(0);
-    opacity: 0.9;
-  }
-}
-
-.modal-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: rgba(0, 0, 0, 0.7);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 9999;
-}
-
 .modal-content {
-  position: relative;
-  background: #1a1a1a;
-  padding: 20px;
-  width: 90%;
-  max-width: 400px;
-  margin: auto;
-  @include yellowGreenGradientBorder;
-
   h3 {
     margin-top: 5px;
-    color: #aaa;
     display: flex;
     align-items: center;
     gap: 8px;
@@ -54,11 +21,6 @@ $button-padding-vertical: 8px;
       filter: drop-shadow(0 0 5px rgba(0, 0, 0, 0.8));
       @include peekUpAnimation;
     }
-  }
-
-  p {
-    color: #ccc;
-    line-height: 1.2;
   }
 
   .player {
