+++ src/components/Modal/modal.scss
@@ -1,0 +1,34 @@
+@import '../../styles/gradients.scss';
+
+.modal-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: rgba(0, 0, 0, 0.7);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 9999;
+}
+
+.modal-content {
+  position: relative;
+  background: #1a1a1a;
+  padding: 20px;
+  width: 90%;
+  max-width: 400px;
+  margin: auto;
+  @include yellowGreenGradientBorder;
+
+  h3 {
+    color: #ccc;
+    margin-top: 0;
+  }
+
+  p {
+    color: #aaa;
+    line-height: 1.4;
+  }
+}
