+++ src/components/OpenSourceInfo/index.scss
@@ -37,7 +37,7 @@ $transition-duration: 0.5s;
     @media (max-width: $breakpoint-tablet) {
       position: absolute;
       right: 25px;
-      bottom: 30px;
+      bottom: 25px;
       cursor: pointer;
     }
   }
