+++ src/styles/_animations.scss
@@ -108,7 +108,7 @@
 
 @keyframes peekUp {
   from {
-    transform: translateY(20%) scale(0.6);
+    transform: translateY(50%) scale(0.5);
     opacity: 0;
   }
   to {
